+++ src/main/scala/io/picnicml/doddlemodel/modelselection/HyperparameterSearch.scala
@@ -1,7 +1,7 @@
 package io.picnicml.doddlemodel.modelselection
 
 import cats.syntax.option._
-import io.picnicml.doddlemodel.CrossScalaCompat.doubleOrdering
+import io.picnicml.doddlemodel.CrossScalaCompat.floatOrdering
 import io.picnicml.doddlemodel.data.{Features, IntVector, Target}
 import io.picnicml.doddlemodel.typeclasses.Predictor
 
@@ -29,7 +29,7 @@ class HyperparameterSearch private (val numIterations: Int, val crossVal: CrossV
   def bestOf[A](x: Features, y: Target, groups: Option[IntVector] = none)(generatePredictor: => A)
                (implicit ev: Predictor[A], rand: Random = new Random()): A = {
 
-    case class PredictorWithScore(predictor: A, score: Double)
+    case class PredictorWithScore(predictor: A, score: Float)
 
     val scoresPredictors = (0 until this.numIterations).map { _ =>
       val predictor = generatePredictor
